Initial beta test version

+++ src/main/java/com/service/CategoriesService.java
@@ -3,6 +3,7 @@ package com.service;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.support.PagedListHolder;
 import org.springframework.stereotype.Service;
 
 import com.dao.CategoriesDao;
@@ -33,7 +34,9 @@ public class CategoriesService implements ICategories {
 
 	@Override
 	public void delete(Categories categories) {
-		categoriesDao.delete(categories);
+		if (categories.isProtect() == false) {
+			categoriesDao.delete(categories);
+		}
 	}
 	public List<String> retrieveSubCategories(int household_id, String parent) {
 		return categoriesDao.retrieveSubCategories(household_id, parent);
@@ -42,10 +45,16 @@ public class CategoriesService implements ICategories {
 		return categoriesDao.retrieveCategoryByName(household_id, category);
 	}
 
-	public List<Categories> retrieveList(int household_id, String parent) {
-		return categoriesDao.retrieveList(household_id, parent);
+	public PagedListHolder<Categories> retrieveList(int household_id, String parent) {
+		return new PagedListHolder<Categories>(categoriesDao.retrieveList(household_id, parent));
 	}
 	public boolean exists(int household_id, String category) {
 		return categoriesDao.exists(household_id, category);
+	}
+	public boolean hasSubCategories(int household_id, String category) {
+		return categoriesDao.hasSubCateories(household_id, category);
+	}
+	public List<Categories> retrieveByParent(int household_id, String parent) {
+		return categoriesDao.retrieveByParent(household_id, parent);
 	}
 }
